+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.1.1 (Unreleased)
 
+  * Fix installation path/unzip on windows
   * Add wrapper script to address zombie processes
 
 ## v0.1.0 (2021-07-25)